Added metrics

+++ cnn-deblur/model/deblur_gan.py
@@ -127,9 +127,11 @@ class DeblurGan:
                 self.discriminator.trainable = False
                 c_loss = self.combined.train_on_batch(blur_batch, [sharp_batch, output_true_batch])
                 c_losses.append(c_loss)
-                psnr_metric = psnr(tf.convert_to_tensor(sharp_batch), tf.convert_to_tensor(generated_batch))
+                psnr_metric = psnr(tf.convert_to_tensor(sharp_batch, dtype=tf.double),
+                                   tf.convert_to_tensor(generated_batch, dtype=tf.double))
                 psnr_metrics.append(psnr_metric)
-                ssim_metric = ssim(tf.convert_to_tensor(sharp_batch), tf.convert_to_tensor(generated_batch))
+                ssim_metric = ssim(tf.convert_to_tensor(sharp_batch, dtype=tf.double),
+                                   tf.convert_to_tensor(generated_batch, dtype=tf.double))
                 ssim_metrics.append(ssim_metric)
                 self.discriminator.trainable = True
